Boot 584 - validating body; fixing minor issues
