+++ MyMusic/src/main/java/com/ciandt/summit/bootcamp2022/common/exception/dto/ValidationErrorWithFieldsDto.java
@@ -3,15 +3,17 @@ package com.ciandt.summit.bootcamp2022.common.exception.dto;
 import lombok.Getter;
 
 import java.time.LocalDateTime;
+import java.util.List;
 
 @Getter
 public class ValidationErrorWithFieldsDto extends ValidationErrorDto {
 
-    private String field;
+    private List<String> errors;
 
     public ValidationErrorWithFieldsDto(LocalDateTime timestamp, Integer status,
-                                        String error, String message, String path, String field) {
+                                        String error, String message, String path,
+                                        List<String> errors) {
         super(timestamp, status, error, message, path);
-        this.field = field;
+        this.errors = errors;
     }
 }
