+++ MyMusic/src/main/java/com/ciandt/summit/bootcamp2022/controller/PlaylistController.java
@@ -7,6 +7,7 @@ import com.ciandt.summit.bootcamp2022.service.PlaylistService;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.responses.ApiResponse;
 import io.swagger.v3.oas.annotations.responses.ApiResponses;
+import io.swagger.v3.oas.annotations.tags.Tag;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.log4j.Log4j2;
 import org.springframework.http.HttpStatus;
@@ -20,6 +21,7 @@ import java.util.Set;
 @RestController
 @RequiredArgsConstructor
 @RequestMapping("/api/playlists")
+@Tag(name = "Playlist")
 @Log4j2
 public class PlaylistController {
