+++ MyMusic/src/main/java/com/ciandt/summit/bootcamp2022/service/MusicService.java
@@ -7,6 +7,7 @@ import java.util.Set;
 public interface MusicService {
 
     boolean isAValidSearch(String filter);
+
     Set<Music> searchMusicsByFilter(String filter);
 
     Set<Music> searchAllMusics();
