+++ MyMusic/src/main/resources/application.properties
@@ -39,3 +39,4 @@ spring.output.ansi.enabled=always
 
 # Receive JSON empty string as null object for validation purposes
 spring.jackson.deserialization.ACCEPT_EMPTY_STRING_AS_NULL_OBJECT=true
+spring.jackson.deserialization.fail-on-unknown-properties=true
